% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{ff}
\alias{ff}
\title{Find Canonical Matrix}
\usage{
ff(A, return_p = FALSE, debug = FALSE)
}
\arguments{
\item{A}{a 2x2 integer matrix}

\item{return_p}{Boolean whether to return P such that return value is P^{-1} AP}

\item{debug}{Boolean}
}
\description{
Find Canonical Matrix
}
