% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parallelogram.R
\name{plot_image}
\alias{plot_image}
\title{Plot Image}
\usage{
plot_image(A, dd, det = FALSE)
}
\arguments{
\item{A}{2x2 integer matrix}

\item{dd}{data frame with names x and y containing two points of parallelogram}

\item{det}{Boolean changes the form of plot by printing determinant}
}
\description{
Plot Image
}
