% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parallelogram.R
\name{transform_eigenvectors}
\alias{transform_eigenvectors}
\title{Eigenvector Transform}
\usage{
transform_eigenvectors(A)
}
\arguments{
\item{A}{2x2 integer matrix with comples eigenvalues}
}
\description{
This is only useful to me for testing hypotheses about matrices with complex eigenvalues.
}
